+++ src/DfudsTrie.cpp
@@ -10,8 +10,8 @@ void DfudsTrieLCP::read(istream &is) {
   //display();
 }
 
-uint64_t DfudsTrieLCP::findLCP(const char *key, uint8_t *lcp) const {
-  uint64_t offset = 1;
+uint32_t DfudsTrieLCP::findLCP(const char *key, uint8_t *lcp) const {
+  uint32_t offset = 1;
   if (lcp && *lcp) {
     offset = _offsets[*lcp - 1];
     key += *lcp;
@@ -24,8 +24,8 @@ uint64_t DfudsTrieLCP::findLCP(const char *key, uint8_t *lcp) const {
   return _keyRank(offset);
 }
 
-uint64_t DfudsTrieLCP::rightNearFindLCP(const char *key, uint8_t *lcp) const {
-  uint64_t offset = 1;
+uint32_t DfudsTrieLCP::rightNearFindLCP(const char *key, uint8_t *lcp) const {
+  uint32_t offset = 1;
   if (lcp && *lcp) {
     offset = _offsets[*lcp - 1];
     key += *lcp;
@@ -49,28 +49,28 @@ uint64_t DfudsTrieLCP::rightNearFindLCP(const char *key, uint8_t *lcp) const {
   }
   if (lcp)  *lcp += keyLen;
 
-  uint64_t rank = _keyRank(offset - 1) + 1;
+  uint32_t rank = _keyRank(offset - 1) + 1;
   return rank;
 }
 
-uint64_t DfudsTrie::findLCP(const char *key, uint8_t *lcp) const {
+uint32_t DfudsTrie::findLCP(const char *key, uint8_t *lcp) const {
   return find(key);
 }
 
-uint64_t DfudsTrie::rightNearFindLCP(const char *key, uint8_t *lcp) const {
+uint32_t DfudsTrie::rightNearFindLCP(const char *key, uint8_t *lcp) const {
   return rightNearFind(key);
 }
 
-uint64_t DfudsTrie::find(const char *key) const {
-  uint64_t offset = 1;
+uint32_t DfudsTrie::find(const char *key) const {
+  uint32_t offset = 1;
   bool found = _followKey(key, offset, NULL);
   if (!found)  return 0;
 
   return _keyRank(offset);
 }
 
-uint64_t DfudsTrie::rightNearFind(const char *key) const {
-  uint64_t offset = 1;
+uint32_t DfudsTrie::rightNearFind(const char *key) const {
+  uint32_t offset = 1;
   int keyLen = 0;
   bool found = _followKey(key, offset, &keyLen);
   if (!found) {
@@ -87,20 +87,20 @@ uint64_t DfudsTrie::rightNearFind(const char *key) const {
     } 
   }
 
-  uint64_t rank = _keyRank(offset - 1) + 1;
+  uint32_t rank = _keyRank(offset - 1) + 1;
   return rank;
 }
 
-void DfudsTrie::select(uint64_t rank, string &key) const {
+void DfudsTrie::select(uint32_t rank, string &key) const {
   assert(0 < rank && rank <= _is_keys.rank1(_is_keys.count() - 1));
   key.clear();
   // + 1 to turn offset to rank, -1 to move to the node before.
-  uint64_t nodeOffset = _is_keys.select1Naive(rank) + 1 - 1;
+  uint32_t nodeOffset = _is_keys.select1Naive(rank) + 1 - 1;
   // + 1 to move to the beginning of next node.
-  uint64_t id = _dfuds.select1Naive(nodeOffset) + 1;
+  uint32_t id = _dfuds.select1Naive(nodeOffset) + 1;
 
   while (id != 1) {
-    uint64_t parent = _parent(id);
+    uint32_t parent = _parent(id);
     uint8_t label = _label(parent, id);
     key = (char)label + key;
     id = parent;
@@ -167,9 +167,9 @@ void DfudsTrie::computeOffsets(const string &key,
 
 // id: inout type. 
 bool DfudsTrie::_followKey(const char *key, 
-    uint64_t &id, int *prefixLen, int *depth) const {
+    uint32_t &id, int *prefixLen, int *depth) const {
   int keyLen = strlen(key);
-  uint64_t offset = id;
+  uint32_t offset = id;
   int ki = 0;
   int last = 0;
   for (; ki < keyLen; ki++) {
@@ -186,11 +186,11 @@ bool DfudsTrie::_followKey(const char *key,
   return ki == keyLen;
 }
 
-uint8_t DfudsTrie::_childLowerBound(uint64_t id, uint8_t ch) const {
+uint8_t DfudsTrie::_childLowerBound(uint32_t id, uint8_t ch) const {
   if (_dfuds[id])  return 0;
 
-  uint64_t rankOpen = _dfuds.rank0(id) - 1;
-  uint64_t base = rankOpen - 1;
+  uint32_t rankOpen = _dfuds.rank0(id) - 1;
+  uint32_t base = rankOpen - 1;
 
   for (int i = 0; !_dfuds[id+i]; i++) {
     uint8_t label = _labels[base + i];
@@ -199,11 +199,11 @@ uint8_t DfudsTrie::_childLowerBound(uint64_t id, uint8_t ch) const {
   return 0;
 }
 
-uint8_t DfudsTrie::_childRank(uint64_t id, uint8_t ch) const {
+uint8_t DfudsTrie::_childRank(uint32_t id, uint8_t ch) const {
   if (_dfuds[id])  return 0;
 
-  uint64_t rankOpen = _dfuds.rank0(id) - 1;
-  uint64_t base = rankOpen - 1;
+  uint32_t rankOpen = _dfuds.rank0(id) - 1;
+  uint32_t base = rankOpen - 1;
 
   for (int i = 0; !_dfuds.bitAt(id+i); ++i) {
     uint8_t label = _labels[base + i];
@@ -212,15 +212,15 @@ uint8_t DfudsTrie::_childRank(uint64_t id, uint8_t ch) const {
   return 0;
 }
 
-uint64_t DfudsTrie::_childSelect(uint64_t id, uint8_t rank) const {
-  uint64_t offsetClose = id + _degree(id);
-  uint64_t tmpOffset = offsetClose - rank;
-  uint64_t nextOffset = _dfuds.findClose(tmpOffset) + 1;
+uint32_t DfudsTrie::_childSelect(uint32_t id, uint8_t rank) const {
+  uint32_t offsetClose = id + _degree(id);
+  uint32_t tmpOffset = offsetClose - rank;
+  uint32_t nextOffset = _dfuds.findClose(tmpOffset) + 1;
   return nextOffset;
 }
 
-uint64_t DfudsTrie::_generalizedSibling(uint64_t id) const {
-  uint64_t offset = id;
+uint32_t DfudsTrie::_generalizedSibling(uint32_t id) const {
+  uint32_t offset = id;
   while (true) {
     uint8_t degree = _degree(offset);
     if (degree == 0) break;
@@ -232,34 +232,34 @@ uint64_t DfudsTrie::_generalizedSibling(uint64_t id) const {
   return offset;
 }
 
-uint8_t DfudsTrie::_degree(uint64_t id) const {
+uint8_t DfudsTrie::_degree(uint32_t id) const {
   int offset = id;
   while (!_dfuds[offset])  ++offset;
   uint8_t degree = offset - id;
   return degree;
 }
 
-uint8_t DfudsTrie::_label(uint64_t parent, uint64_t child) const {
-  uint64_t openOffset = _dfuds.findOpenNaive(child - 1);
+uint8_t DfudsTrie::_label(uint32_t parent, uint32_t child) const {
+  uint32_t openOffset = _dfuds.findOpenNaive(child - 1);
   int degree = _degree(parent);
   uint8_t childRank = parent + degree - openOffset;
 
   // -1 for the head open, -1 to turn rank to offset.
-  uint64_t base = _dfuds.rank0(parent) - 1 - 1;
+  uint32_t base = _dfuds.rank0(parent) - 1 - 1;
   return _labels[base + childRank - 1];
 }
 
-uint64_t DfudsTrie::_parent(uint64_t id) const {
-  uint64_t openOffset = _dfuds.findOpenNaive(id - 1);
-  uint64_t closeRank = _dfuds.rank1(openOffset);
+uint32_t DfudsTrie::_parent(uint32_t id) const {
+  uint32_t openOffset = _dfuds.findOpenNaive(id - 1);
+  uint32_t closeRank = _dfuds.rank1(openOffset);
   if (closeRank == 0)  return 1;
-  uint64_t result = _dfuds.select1Naive(closeRank) + 1;
+  uint32_t result = _dfuds.select1Naive(closeRank) + 1;
   return result;
 }
 
-uint64_t DfudsTrie::_keyRank(uint64_t id) const {
-  uint64_t close_offset = id + _degree(id);
-  uint64_t key_offset = _dfuds.rank1(close_offset) - 1;
-  uint64_t rank = _is_keys.rank1(key_offset);
+uint32_t DfudsTrie::_keyRank(uint32_t id) const {
+  uint32_t close_offset = id + _degree(id);
+  uint32_t key_offset = _dfuds.rank1(close_offset) - 1;
+  uint32_t rank = _is_keys.rank1(key_offset);
   return rank;
 }
